requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
google-api-python-client>=2.0.0
pandas>=2.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
requests>=2.31.0
streamlit-qrcode-scanner>=0.0.4
